requirements.txt
fastapi
uvicorn[standard]
pinecone
google-generativeai
python-dotenv
PyPDF2
python-docx
requests
